The Great Cleanup of 12 March
I organized in the order of:
@FXML before nonFXML
public->no prefix->private
levels of abstraction
multilined methods->single lines (or getters/setters)

Also, I made everything private that I did not see a reason to be
public(excluding some ServerThread code I wasn't comfortable enough with
to know if it would be ok to make private). If any of these methods need
to be public, feel free to change them back.
Also also, @FXML only needs to prefix codes that are explicitly
referenced by an element in Scenebuilder.